backend/requirements.txt
langchain
langchain-community
pypdf
tiktoken
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-core
black
isort
ruff
pre-commit
mypy
pytest
pylint
python-dotenv
fastapi
sqlalchemy
sqlalchemy.orm
passlib
datetime
python-jose
pydantic
pydantic[email]